make prompt by using multithreadand multiprocess
