+++ client/Makefile
@@ -11,7 +11,6 @@ CFLAGS = -Wall -Wextra -Werror -O3
 SRCS = ./client.c\
        ./prompt.c\
        ./connector.c\
-       ./thread.c\
 
 OBJS = $(SRCS:.c=.o)
