+++ server/setting.h
@@ -3,5 +3,6 @@
 
 # define LISTEN_QUEUE_SIZE 5
 # define THREAD_POOL_SIZE 10
+# define BUFFER_SIZE 100
 
 #endif
